requirements-prod.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
gemini-webapi>=1.17.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
